Added awareness about being protectable.
Basically, this is a rather poor way of preventing
the actual before hook methods from fulfilling their
purpose, in case the current Resource is not protectable.

I think it would be nicer, if Extlib::Hook would provide
a way of specifiying a condition that would be evaluated
before it decides to call the advised, or the original
method.